requirements.txt
pillow==12.0.0
pygame==2.6.1
setuptools==80.9.0
wheel==0.45.1
sounddevice>=0.5.3
numpy==2.3.4
scipy>=1.11.0
openai==2.6.1
python-dotenv==1.0.1
opencv-python>=4.8.0
# Optional: enable if you want SVG -> raster conversion at runtime
cairosvg==2.7.1
PyYAML==6.0.3
